requirements.txt
google-generativeai>=0.3.2
flask==3.0.3
python-dotenv==1.0.0
flask-cors==4.0.0
fpdf2==2.7.7
python-pptx==0.6.23